# REF: Move sindex to GeometryArray