--- a/geopandas/array.py
+++ b/geopandas/array.py
@@ -285,7 +285,10 @@ class GeometryArray(ExtensionArray):
 
     @property
     def sindex(self):
-        if self._sindex is not None:  # check for existing
+        if self._sindex is None:
+            self._sindex = get_sindex_class()(self.data)
+            
+        return self._sindex
             pass
         else:  # build it
             self._sindex = get_sindex_class()(self.data)