--- a/geopandas/array.py
+++ b/geopandas/array.py
@@ -287,7 +287,6 @@ class GeometryArray(ExtensionArray):
     def sindex(self):
         if self._sindex is None:
             self._sindex = get_sindex_class()(self.data) 
-
         return self._sindex
 
     @property