Crossref #1439.

This does the following (roughly steps 1 & 2 in #1439):
1. Drop the `objects` parameter completely.
2. Move the management of the spatial index to `GeometryArray`.

I don't expect this to be merged until after 0.8.0, but I'm submitting for some preliminary feedback.