@adriangb 2020-07-06
@martinfleis is now a good time to work on this? I can resolve conflicts and throw it back up for review. Thanks!