@martinfleis 2020-07-06
@adriangb yeah, can't see why not, go ahead.