@adriangb 2020-07-08
Ok all of the unnecessary changes you caught are reverted. I also opened #1508 to deal with `base.HAS_SINDEX`.