@adriangb 2020-07-19
Waiting to run benchmarks. Do you think we should add a public method to check if sindex has been generated?