@martinfleis 2020-07-19
In the light of #1523, I would say it could be useful. But feel free to leave it as a follow up PR.