@adriangb 2020-07-19
👍

On Sun, Jul 19, 2020, 7:43 AM Martin Fleischmann <notifications@github.com>
wrote:

> In the light of #1523 <https://github.com/geopandas/geopandas/issues/1523>,
> I would say it could be useful. But feel free to leave it as a follow up PR.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/geopandas/geopandas/pull/1444#issuecomment-660638197>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AANMPPY3DD43D3CX24SQH7TR4LTAPANCNFSM4NIQW6JA>
> .
>