@adriangb 2020-07-19
So I'm seeing some very strange benchmark results. Some are coming in at x3 and some at x0.1! @martinfleis  could you take a look at the [attached benchmarks](https://github.com/geopandas/geopandas/files/4944086/comp.txt), and maybe re-run if you have time? I tried to run cProfile to see if the sindex's `__init__` was being called multiple times or something like that, but everything looked okay. I can't imagine anything in these changes would really impact performance (aside from a couple extra `__getattr__` calls or something like that).