@martinfleis 2020-07-20
I have no idea why, but also see some significant regressions and few significant speedups. But some of those regressions are up to 350x slower than on master. However, not all coincide with your results... There seems to be some instability of performance I do not understand.