@adriangb 2020-07-20
Could this be from creation/destruction of GeometryArrays? I don't see where or how that would happen, I'm just guessing...

To me the most bizarre part is that there is a 100x speedup for some benchmarks 🤨!

Good call on running the benchmarks by the way.