@martinfleis 2020-07-20
Is this still true? Is it possible that there's some caching happening? Or was and is not anymore? In your timings, the issue is mostly with pygeos.

https://github.com/geopandas/geopandas/blob/d3228ddb6814515506921ad5c2e6d6f8bd99c9ae/benchmarks/sindex.py#L35-L36