@adriangb 2020-07-20
> Is this still true? Is it possible that there's some caching happening? Or was and is not anymore?

I'm not 100% sure to be honest, I only added that comment based on @jorisvandenbossche 's suggestion that that was the case (which I may have misinterpreted). In any case, I don't know that that would explain the differences in `query` or `query_bulk`.