@adriangb 2020-07-20
I just looked through the changes in `sindex.py` and ran cProfile on just that part, I do not see any differences (as expected):
```python3
try:
    # for this branch
    myindex = sindex.PyGEOSSTRTreeIndex(mixed.geometry.values.data)
except AttributeError:
    # for master
    myindex = sindex.PyGEOSSTRTreeIndex(mixed.geometry)

data = mixed.geometry.values.data

def bench():
    myindex.query_bulk(data)

cProfile.run(statement='bench()', sort="cumtime")
```