@martinfleis 2020-07-20
@brendan-ward do you have any idea what could be happening here?