@brendan-ward 2020-07-20
> pygeos will only create the index once

This comes from [GEOS](https://geos.osgeo.org/doxygen/classgeos_1_1index_1_1strtree_1_1STRtree.html); the tree is not built until the first query.

The first call to `query()` or `query_bulk()` will incur the cost of building the tree, not the call to initialize it.  In the context of benchmarks, if there are repeated `query` / `query_bulk` calls to the same tree instance, the first will be slow, the remainder should be faster.  This may throw off the average but should be most obvious in the `+/-` values (wider variation).  I'm not immediately seeing that in the benchmarks here.

I haven't traced it all the way through, but it looks like this change in the benchmarks:
```
self.data[tree_geom_type].geometry.values._sindex = None
self.data[tree_geom_type].sindex
```

may be forcing recreation of the pygeos STRtree in this PR compared to the tree being reused in `master`.  If that is the case, I'd expect the `query()` / `query_bulk()` benchmarks here to be slower since this would incur the cost of creating the tree * benchmark repeats.

However, I would not expect a significant change in the performance of index creation in `sindex.Bench.time_index_creation`.  That seems like it may be related to how the geometries are prepared before passing in to the tree.

Is it possible that the changes here are causing geometries to go through `shapely` first [as per this line](https://github.com/geopandas/geopandas/pull/1444/files#diff-ab558acce76aadad9a6552df487654baR335-R336)?  


Unless I'm reading it incorrectly, it looks like the biggest differences in performance here are for `rtree`, which smaller ones for `pygeos`?