@adriangb 2020-07-21
Thanks for the input, that makes sense. I re-structured the benchmarks to pre-build the index _and_ make a call to `query` to make 100% sure it's built before benchmarking. The results (attachment) still don't make much sense to me.

> Is it possible that the changes here are causing geometries to go through shapely first as per this line?

@brendan-ward can you re-link that line you are referring to, the link didn't seem to work well for me.