@brendan-ward 2020-07-21
@adriangb `sindex.py` around line 354:

```
if isinstance(geometry, BaseGeometry):
    geometry = array._shapely_to_geom(geometry)
```

Nothing in these changes or prior `sindex` changes implies that geometries would go through `shapely` unnecessarily, but where there seems to be dramatic slowdowns related to passing around arrays of geometries, it is something I'd want to rule out to be sure.