@adriangb 2020-07-21
Just to make sure, I added a print statement to `_shapely_to_geom`  _and_ `_geom_to_shapely` which helped me catch an extra conversion happening in the new `time_query` benchmark, but after fixing that I did not see any more conversions in the rest of the benchmarks (except when calculating the bounds for `time_intersects`, but that's a necessary conversion). I'm re-running the benchmarks now.