@brendan-ward 2020-07-22
Whoops - I hadn't checked into how the test data were being generated and obviously forgot about that from prior PRs.  The data are on the whole deterministic; the only place randomness would affect results is in `query()` since it is drawing a single geometry from the test geometries randomly.

`query_bulk()` should be deterministic because it is using all geometries of a given deterministic collection.

I worked up an alternative set of benchmarks using `pytest-benchmark` to see if it produces more deterministic results.
[sindex.py.txt](https://github.com/geopandas/geopandas/files/4957779/sindex.py.txt)


In results below, first part of the benchmark name is tree geometry, query geometry.

Apologies for the changes in units and change in sort of the individual benchmarks, those are automatically determined.

In general, it looks like across the board, the benchmarks are slower in this PR than compared to `master` (unless I did something wrong in constructing the benchmarks).  I'm not immediately seeing the same nondeterministic behavior as in the `asv` based benchmarks, but I need to run these again on the same branch to investigate that further.


**Pygeos index creation:**

`master`:
```-------------------------------------------------------------------------------------------- benchmark 'create': 3 tests ---------------------------------------------------------------------------------------------
Name (time in ns)                             Min                    Max                Mean              StdDev              Median               IQR            Outliers  OPS (Mops/s)            Rounds  Iterations
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_create[points-tree_geometry1]       282.5000 (1.0)      16,564.9000 (1.0)      304.2919 (1.0)      204.5769 (1.09)     288.5000 (1.0)      4.0000 (1.05)      144;520        3.2863 (1.0)       10000          10
test_create[polygons-tree_geometry0]     288.6000 (1.02)     30,450.8000 (1.84)     307.5033 (1.01)     317.2916 (1.69)     295.1000 (1.02)     4.3000 (1.13)       62;410        3.2520 (0.99)      10000          10
test_create[mixed-tree_geometry2]        290.5000 (1.03)     16,947.3000 (1.02)     304.6672 (1.00)     187.3634 (1.0)      296.7000 (1.03)     3.8000 (1.0)        73;291        3.2823 (1.00)      10000          10
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

this PR (note change in units):
```
--------------------------------------------------------------------------------------- benchmark 'create': 3 tests ---------------------------------------------------------------------------------------
Name (time in us)                           Min                 Max              Mean            StdDev            Median               IQR            Outliers  OPS (Kops/s)            Rounds  Iterations
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_create[polygons-tree_geometry0]     5.2873 (1.0)      160.5450 (7.22)     6.4524 (1.13)     3.3089 (3.28)     5.5164 (1.01)     0.9397 (11.02)    771;1166      154.9806 (0.88)      10000          10
test_create[mixed-tree_geometry2]        5.2912 (1.00)      26.6711 (1.20)     5.6855 (1.0)      1.0086 (1.0)      5.4575 (1.00)     0.0853 (1.0)      521;1168      175.8858 (1.0)       10000          10
test_create[points-tree_geometry1]       5.2925 (1.00)      22.2447 (1.0)      5.7800 (1.02)     1.2408 (1.23)     5.4564 (1.0)      0.0870 (1.02)     666;1398      173.0097 (0.98)      10000          10
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

This has me a bit stumped; it should be very fast in both cases and I did not expect variation in creation time between 'master` and this PR.


**Pygeos index query**
Instead of sampling a geometry to use for query, this iterates over every geometry a given set of geometries.  The idea being that this averages out individual performance differences between geometries, but does incur some overhead for looping through them (so don't compare these to rtree)

`master`
```
----------------------------------------------------------------------------------------------------------- benchmark 'query': 9 tests ----------------------------------------------------------------------------------------------------------
Name (time in us)                                                       Min                   Max                  Mean             StdDev                Median                IQR            Outliers         OPS            Rounds  Iterations
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_query[points,points-tree_geometry4-query_geometry4]           294.3290 (1.0)        450.4478 (1.0)        306.0421 (1.0)      16.8538 (1.18)       300.3525 (1.0)      10.6744 (1.0)         21;17  3,267.5241 (1.0)         250           5
test_query[points,polygons-tree_geometry3-query_geometry3]         321.9102 (1.09)       512.1192 (1.14)       338.7333 (1.11)     23.6080 (1.65)       330.0957 (1.10)     14.4362 (1.35)        24;22  2,952.1749 (0.90)        250           5
test_query[polygons,points-tree_geometry1-query_geometry1]         407.3762 (1.38)       655.8082 (1.46)       429.5085 (1.40)     29.9674 (2.10)       419.9938 (1.40)     22.1990 (2.08)        22;18  2,328.2424 (0.71)        250           5
test_query[polygons,polygons-tree_geometry0-query_geometry0]       462.7316 (1.57)       612.5112 (1.36)       485.6371 (1.59)     23.5121 (1.65)       478.2715 (1.59)     18.4092 (1.72)        22;18  2,059.1507 (0.63)        250           5
test_query[mixed,points-tree_geometry7-query_geometry7]            518.7800 (1.76)       611.1120 (1.36)       536.2729 (1.75)     14.2829 (1.0)        531.6322 (1.77)     14.5672 (1.36)        46;17  1,864.7221 (0.57)        250           5
test_query[mixed,polygons-tree_geometry6-query_geometry6]          604.3654 (2.05)       775.0392 (1.72)       632.9642 (2.07)     22.2991 (1.56)       627.7125 (2.09)     20.8418 (1.95)        34;13  1,579.8682 (0.48)        250           5
test_query[points,mixed-tree_geometry5-query_geometry5]            615.3478 (2.09)       966.2834 (2.15)       641.0579 (2.09)     35.0780 (2.46)       629.5927 (2.10)     19.8992 (1.86)        20;21  1,559.9216 (0.48)        250           5
test_query[polygons,mixed-tree_geometry2-query_geometry2]          859.9612 (2.92)     1,040.9270 (2.31)       896.5372 (2.93)     29.2314 (2.05)       887.7495 (2.96)     21.4014 (2.00)        27;18  1,115.4027 (0.34)        250           5
test_query[mixed,mixed-tree_geometry8-query_geometry8]           1,119.6912 (3.80)     1,356.9206 (3.01)     1,164.5867 (3.81)     34.7289 (2.43)     1,154.3654 (3.84)     27.4622 (2.57)        36;17    858.6737 (0.26)        250           5
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

This PR (note change in units)

```
------------------------------------------------------------------------------------------------- benchmark 'query': 9 tests ------------------------------------------------------------------------------------------------
Name (time in ms)                                                   Min               Max              Mean            StdDev            Median               IQR            Outliers       OPS            Rounds  Iterations
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_query[points,polygons-tree_geometry3-query_geometry3]       1.3151 (1.0)      6.7100 (3.40)     1.5727 (1.03)     0.5021 (6.56)     1.4241 (1.0)      0.1904 (2.58)        15;25  635.8491 (0.97)        250           5
test_query[points,points-tree_geometry4-query_geometry4]         1.4094 (1.07)     2.1026 (1.07)     1.5332 (1.0)      0.1142 (1.49)     1.4974 (1.05)     0.0796 (1.08)        33;29  652.2286 (1.0)         250           5
test_query[polygons,polygons-tree_geometry0-query_geometry0]     1.4787 (1.12)     1.9729 (1.0)      1.5843 (1.03)     0.0766 (1.0)      1.5642 (1.10)     0.0738 (1.0)         54;14  631.1887 (0.97)        250           5
test_query[polygons,points-tree_geometry1-query_geometry1]       1.5665 (1.19)     2.2030 (1.12)     1.7066 (1.11)     0.1078 (1.41)     1.6744 (1.18)     0.1082 (1.47)        56;15  585.9508 (0.90)        250           5
test_query[mixed,polygons-tree_geometry6-query_geometry6]        1.6667 (1.27)     3.2801 (1.66)     1.8981 (1.24)     0.2297 (3.00)     1.8099 (1.27)     0.2467 (3.34)        30;11  526.8315 (0.81)        250           5
test_query[mixed,points-tree_geometry7-query_geometry7]          1.7037 (1.30)     2.3013 (1.17)     1.8163 (1.18)     0.0938 (1.23)     1.7901 (1.26)     0.0747 (1.01)        42;25  550.5749 (0.84)        250           5
test_query[points,mixed-tree_geometry5-query_geometry5]          2.7489 (2.09)     9.6544 (4.89)     3.3668 (2.20)     0.8282 (10.82)    3.1065 (2.18)     0.5836 (7.91)        18;16  297.0155 (0.46)        250           5
test_query[polygons,mixed-tree_geometry2-query_geometry2]        3.1021 (2.36)     7.7435 (3.92)     3.7576 (2.45)     0.7451 (9.73)     3.3671 (2.36)     0.9250 (12.54)        33;6  266.1251 (0.41)        250           5
test_query[mixed,mixed-tree_geometry8-query_geometry8]           3.4130 (2.60)     5.0239 (2.55)     3.6486 (2.38)     0.2730 (3.57)     3.5458 (2.49)     0.1694 (2.30)        32;32  274.0746 (0.42)        250           5
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

**Same as above, but using `predicate="intersects"`**

`master`
```
--------------------------------------------------------------------------------------------------------------- benchmark 'query-intersects': 9 tests ----------------------------------------------------------------------------------------------------------------
Name (time in us)                                                                   Min                    Max                   Mean                StdDev                 Median                   IQR            Outliers         OPS            Rounds  Iterations
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_query_intersects[points,points-tree_geometry4-query_geometry4]            574.4708 (1.0)       1,198.6454 (1.0)         668.1895 (1.0)         86.4568 (1.42)        635.3519 (1.0)         87.1006 (2.18)        42;14  1,496.5815 (1.0)         250           5
test_query_intersects[polygons,points-tree_geometry1-query_geometry1]        1,150.8710 (2.00)      3,850.1848 (3.21)      1,700.8372 (2.55)       446.7708 (7.32)      1,653.3610 (2.60)       601.0288 (15.02)        94;3    587.9457 (0.39)        250           5
test_query_intersects[mixed,points-tree_geometry7-query_geometry7]           1,294.3432 (2.25)      1,790.4860 (1.49)      1,355.5407 (2.03)        61.0306 (1.0)       1,338.5903 (2.11)        40.0212 (1.0)         22;18    737.7130 (0.49)        250           5
test_query_intersects[points,polygons-tree_geometry3-query_geometry3]        2,145.0204 (3.73)      4,994.6918 (4.17)      2,700.1724 (4.04)       561.1781 (9.20)      2,444.8221 (3.85)       660.0876 (16.49)       35;11    370.3467 (0.25)        250           5
test_query_intersects[points,mixed-tree_geometry5-query_geometry5]           2,714.9682 (4.73)      5,226.4256 (4.36)      2,935.9834 (4.39)       346.2402 (5.67)      2,825.7948 (4.45)        88.9446 (2.22)        23;36    340.6014 (0.23)        250           5
test_query_intersects[polygons,polygons-tree_geometry0-query_geometry0]     20,990.7078 (36.54)    33,896.7654 (28.28)    22,487.4961 (33.65)    2,187.4126 (35.84)    21,593.4177 (33.99)    1,367.5774 (34.17)       28;28     44.4692 (0.03)        250           5
test_query_intersects[mixed,polygons-tree_geometry6-query_geometry6]        21,197.6732 (36.90)    42,993.0560 (35.87)    23,137.2236 (34.63)    3,410.5487 (55.88)    21,904.8484 (34.48)      857.4950 (21.43)       28;39     43.2204 (0.03)        250           5
test_query_intersects[polygons,mixed-tree_geometry2-query_geometry2]        22,069.2264 (38.42)    39,541.8940 (32.99)    24,021.6801 (35.95)    3,053.8265 (50.04)    22,711.7305 (35.75)    1,154.1000 (28.84)       32;44     41.6291 (0.03)        250           5
test_query_intersects[mixed,mixed-tree_geometry8-query_geometry8]           22,687.5712 (39.49)    28,292.6620 (23.60)    23,272.7915 (34.83)      563.5291 (9.23)     23,148.1953 (36.43)      344.9516 (8.62)        16;14     42.9686 (0.03)        250           5
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

This PR (note change in units)
```
-------------------------------------------------------------------------------------------------- benchmark 'query-intersects': 9 tests ---------------------------------------------------------------------------------------------------
Name (time in ms)                                                               Min                Max               Mean            StdDev             Median               IQR            Outliers       OPS            Rounds  Iterations
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_query_intersects[points,points-tree_geometry4-query_geometry4]          1.7033 (1.0)       2.7937 (1.0)       1.8480 (1.0)      0.1474 (1.78)      1.8041 (1.0)      0.0701 (1.22)        27;32  541.1148 (1.0)         250           5
test_query_intersects[polygons,points-tree_geometry1-query_geometry1]        2.3848 (1.40)      5.3104 (1.90)      2.8496 (1.54)     0.4684 (5.67)      2.6413 (1.46)     0.5027 (8.78)        31;12  350.9248 (0.65)        250           5
test_query_intersects[mixed,points-tree_geometry7-query_geometry7]           2.4883 (1.46)      3.1147 (1.11)      2.6008 (1.41)     0.0827 (1.0)       2.5853 (1.43)     0.0573 (1.0)         27;17  384.4935 (0.71)        250           5
test_query_intersects[points,polygons-tree_geometry3-query_geometry3]        3.2596 (1.91)      5.9713 (2.14)      3.6198 (1.96)     0.4995 (6.04)      3.4065 (1.89)     0.1969 (3.44)        40;42  276.2573 (0.51)        250           5
test_query_intersects[points,mixed-tree_geometry5-query_geometry5]           5.0144 (2.94)      7.4200 (2.66)      5.2280 (2.83)     0.2554 (3.09)      5.1576 (2.86)     0.1030 (1.80)        19;35  191.2765 (0.35)        250           5
test_query_intersects[mixed,polygons-tree_geometry6-query_geometry6]        22.8341 (13.41)    27.5839 (9.87)     23.2327 (12.57)    0.4540 (5.49)     23.1294 (12.82)    0.2961 (5.17)        12;10   43.0428 (0.08)        250           5
test_query_intersects[polygons,polygons-tree_geometry0-query_geometry0]     23.1487 (13.59)    37.9926 (13.60)    25.2255 (13.65)    2.7677 (33.48)    24.0939 (13.35)    1.7689 (30.88)       34;34   39.6424 (0.07)        250           5
test_query_intersects[polygons,mixed-tree_geometry2-query_geometry2]        24.8694 (14.60)    68.2375 (24.43)    26.5352 (14.36)    3.4801 (42.10)    25.5913 (14.18)    1.2413 (21.67)       18;26   37.6858 (0.07)        250           5
test_query_intersects[mixed,mixed-tree_geometry8-query_geometry8]           25.4314 (14.93)    33.8506 (12.12)    25.9112 (14.02)    0.7062 (8.54)     25.7565 (14.28)    0.3232 (5.64)        11;13   38.5933 (0.07)        250           5
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```